SMS Sending added
SMS Sending feature added.
Has to be tested.
Changed contact number's vartype

+++ bloodbank/database/migrations/2016_08_13_190522_create_donors_table.php
@@ -18,7 +18,7 @@ class CreateDonorsTable extends Migration
             $table->string('Password',100);
             $table->string('BloodGroup',10);
             $table->date('LastDonated');
-            $table->integer('Contact');
+            $table->double('Contact');
             $table->string('Address',256);
             $table->timestamps();
         });
